backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
httpx>=0.26.0
websockets>=12.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
mypy>=1.8.0
ruff>=0.1.0
